LLM/EquityResearchReports/backend/stock-viewer-service/requirements.txt
influxdb_client
pandas
flask
flask-cors
python-consul